+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/__init__.py
@@ -18,6 +18,7 @@ from ._models import (
     RepositoryProperties,
     ArtifactTagProperties,
     GetManifestResult,
+    DigestValidationError,
 )
 from ._download_stream import DownloadBlobStream
 from ._version import VERSION
@@ -35,4 +36,5 @@ __all__ = [
     "ArtifactTagProperties",
     "GetManifestResult",
     "DownloadBlobStream",
+    "DigestValidationError",
 ]