+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/job/parameterized_spark.py
@@ -6,7 +6,7 @@
 import re
 from typing import Any, Dict, List
 
-from marshmallow import INCLUDE, ValidationError, fields, post_dump, post_load, pre_dump, pre_load, validate, validates
+from marshmallow import ValidationError, fields, post_dump, post_load, pre_dump, pre_load, validates
 
 from azure.ai.ml._schema.core.fields import CodeField, NestedField
 from azure.ai.ml._schema.core.schema import PathAwareSchema
@@ -53,39 +53,6 @@ CONF_KEY_MAP = {
 }
 
 
-class SparkConfSchema(metaclass=PatchedSchemaMeta):
-    driver_cores = fields.Int()
-    driver_memory = fields.Str(validate=validate.Regexp(re_memory_pattern))
-    executor_cores = fields.Int()
-    executor_memory = fields.Str(validate=validate.Regexp(re_memory_pattern))
-    executor_instances = fields.Int()
-    dynamic_allocation_enabled = fields.Bool()
-    dynamic_allocation_min_executors = fields.Int()
-    dynamic_allocation_max_executors = fields.Int()
-
-    @pre_load
-    def deserialize_field_names(self, data, **kwargs):
-        for field_key, dict_key in CONF_KEY_MAP.items():
-            value = data.get(dict_key, None)
-            if dict_key in data and value is not None:
-                del data[dict_key]
-                data[field_key] = value
-        return data
-
-    @post_dump(pass_original=True)
-    def serialize_field_names(self, data: Dict[str, Any], original_data: Dict[str, Any], **kwargs):
-        # pass original data here to dump conf fields which are not defined in SparkConfSchema
-        for field_name, _ in original_data.items():
-            if field_name not in data:
-                data[field_name] = original_data[field_name]
-        for field_name, dict_name in CONF_KEY_MAP.items():
-            val = data.get(field_name, None)
-            if field_name in data and val is not None:
-                del data[field_name]
-                data[dict_name] = val
-        return data
-
-
 def no_duplicates(name: str, value: List):
     if len(value) != len(set(value)):
         raise ValidationError(f"{name} must not contain duplicate entries.")
@@ -102,7 +69,8 @@ class ParameterizedSparkSchema(PathAwareSchema):
     jars = fields.List(fields.Str(required=True))
     files = fields.List(fields.Str(required=True))
     archives = fields.List(fields.Str(required=True))
-    conf = NestedField(SparkConfSchema, unknown=INCLUDE)
+    conf = fields.Dict(keys=fields.Str(), values=fields.Raw())
+    properties = fields.Dict(keys=fields.Str(), values=fields.Raw())
     environment = UnionField(
         [
             NestedField(AnonymousEnvironmentSchema),
@@ -128,6 +96,37 @@ class ParameterizedSparkSchema(PathAwareSchema):
     @validates("archives")
     def no_duplicate_archives(self, value):
         no_duplicates("archives", value)
+
+    @pre_load
+    def deserialize_field_names(self, data, **kwargs):
+        conf = data["conf"] if "conf" in data else None
+        if conf is not None:
+            for field_key, dict_key in CONF_KEY_MAP.items():
+                value = conf.get(dict_key, None)
+                if dict_key in conf and value is not None:
+                    del conf[dict_key]
+                    conf[field_key] = value
+            data["conf"] = conf
+        return data
+
+    @post_dump(pass_original=True)
+    def serialize_field_names(self, data: Dict[str, Any], original_data: Dict[str, Any], **kwargs):
+        conf = data["conf"] if "conf" in data else {}
+        if conf is not None or original_data.conf is not None:
+            for field_name, value in original_data.conf.items():
+                if field_name not in conf:
+                    if isinstance(value, str) and value.isdigit():
+                        value = int(value)
+                    conf[field_name] = value
+            for field_name, dict_name in CONF_KEY_MAP.items():
+                val = conf.get(field_name, None)
+                if field_name in conf and val is not None:
+                    if isinstance(val, str) and val.isdigit():
+                        val = int(val)
+                    del conf[field_name]
+                    conf[dict_name] = val
+            data["conf"] = conf
+        return data
 
     @post_load
     def demote_conf_fields(self, data, **kwargs):