+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_load_functions.py
@@ -38,6 +38,7 @@ from azure.ai.ml.entities._workspace.connections.workspace_connection import Wor
 from azure.ai.ml.entities._workspace.workspace import Workspace
 from azure.ai.ml.entities._assets._artifacts._package.model_package import ModelPackage
 from azure.ai.ml.exceptions import ErrorCategory, ErrorTarget, ValidationErrorType, ValidationException
+from azure.ai.ml._utils._experimental import experimental
 
 module_logger = logging.getLogger(__name__)
 
@@ -729,7 +730,7 @@ def load_schedule(
     return load_common(Schedule, source, relative_origin, **kwargs)
 
 
-def _load_feature_store(
+def load_feature_store(
     source: Union[str, PathLike, IO[AnyStr]],
     *,
     relative_origin: Optional[str] = None,
@@ -757,7 +758,7 @@ def _load_feature_store(
     return load_common(FeatureStore, source, relative_origin, **kwargs)
 
 
-def _load_feature_set(
+def load_feature_set(
     source: Union[str, PathLike, IO[AnyStr]],
     *,
     relative_origin: Optional[str] = None,
@@ -788,7 +789,7 @@ def _load_feature_set(
     return load_common(FeatureSet, source, relative_origin, **kwargs)
 
 
-def _load_feature_store_entity(
+def load_feature_store_entity(
     source: Union[str, PathLike, IO[AnyStr]],
     *,
     relative_origin: Optional[str] = None,
@@ -819,6 +820,7 @@ def _load_feature_store_entity(
     return load_common(FeatureStoreEntity, source, relative_origin, **kwargs)
 
 
+@experimental
 def load_model_package(
     source: Union[str, PathLike, IO[AnyStr]],
     *,