+++ sdk/servicebus/azure-servicebus/azure/servicebus/_servicebus_client.py
@@ -6,8 +6,7 @@ from typing import Any, Union, Optional, TYPE_CHECKING
 import logging
 from weakref import WeakSet
 from typing_extensions import Literal
-
-import uamqp
+import certifi
 
 from ._base_handler import (
     _parse_conn_str,
@@ -29,6 +28,8 @@ from ._common.constants import (
     ServiceBusSessionFilter,
 )
 
+from ._transport._pyamqp_transport import PyamqpTransport
+
 if TYPE_CHECKING:
     from azure.core.credentials import (
         TokenCredential,
@@ -42,7 +43,7 @@ NextAvailableSessionType = Literal[ServiceBusSessionFilter.NEXT_AVAILABLE]
 _LOGGER = logging.getLogger(__name__)
 
 
-class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-keyword
+class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-keyword,too-many-instance-attributes
     """The ServiceBusClient class defines a high level interface for
     getting ServiceBusSender and ServiceBusReceiver.
 
@@ -84,6 +85,9 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
     :keyword str connection_verify: Path to the custom CA_BUNDLE file of the SSL certificate which is used to
      authenticate the identity of the connection endpoint.
      Default is None in which case `certifi.where()` will be used.
+    :keyword uamqp_transport: Whether to use the `uamqp` library as the underlying transport. The default value is
+     False and the Pure Python AMQP library will be used as the underlying transport.
+    :paramtype uamqp_transport: bool
 
     .. admonition:: Example:
 
@@ -109,50 +113,63 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
         retry_mode: str = "exponential",
         **kwargs: Any
     ) -> None:
+        uamqp_transport = kwargs.pop("uamqp_transport", False)
+        if uamqp_transport:
+            try:
+                from ._transport._uamqp_transport import UamqpTransport
+            except ImportError:
+                raise ValueError("To use the uAMQP transport, please install `uamqp>=1.6.3,<2.0.0`.")
+        self._amqp_transport = UamqpTransport if uamqp_transport else PyamqpTransport
+
         # If the user provided http:// or sb://, let's be polite and strip that.
         self.fully_qualified_namespace = strip_protocol_from_uri(
             fully_qualified_namespace.strip()
         )
 
         self._credential = credential
+        # TODO: can we remove this here? it's recreated in Sender/Receiver
         self._config = Configuration(
             retry_total=retry_total,
             retry_backoff_factor=retry_backoff_factor,
             retry_backoff_max=retry_backoff_max,
             retry_mode=retry_mode,
+            hostname=self.fully_qualified_namespace,
+            amqp_transport=self._amqp_transport,
             **kwargs
         )
         self._connection = None
         # Optional entity name, can be the name of Queue or Topic.  Intentionally not advertised, typically be needed.
         self._entity_name = kwargs.get("entity_name")
-        self._auth_uri = "sb://{}".format(self.fully_qualified_namespace)
+        self._auth_uri = f"sb://{self.fully_qualified_namespace}"
         if self._entity_name:
-            self._auth_uri = "{}/{}".format(self._auth_uri, self._entity_name)
+            self._auth_uri = f"{self._auth_uri}/{self._entity_name}"
         # Internal flag for switching whether to apply connection sharing, pending fix in uamqp library
         self._connection_sharing = False
-        self._handlers = WeakSet()  # type: WeakSet
-
+        self._handlers: WeakSet = WeakSet()
         self._custom_endpoint_address = kwargs.get('custom_endpoint_address')
         self._connection_verify = kwargs.get("connection_verify")
 
     def __enter__(self):
         if self._connection_sharing:
-            self._create_uamqp_connection()
+            self._create_connection()
         return self
 
     def __exit__(self, *args):
         self.close()
 
-    def _create_uamqp_connection(self):
+    def _create_connection(self):
         auth = create_authentication(self)
-        self._connection = uamqp.Connection(
-            hostname=self.fully_qualified_namespace,
-            sasl=auth,
-            debug=self._config.logging_enable,
+        self._connection = self._amqp_transport.create_connection(
+            host=self.fully_qualified_namespace,
+            auth=auth.sasl,
+            network_trace=self._config.logging_enable,
+            custom_endpoint_address=self._custom_endpoint_address,
+            ssl_opts={'ca_certs': self._connection_verify or certifi.where()},
+            transport_type=self._config.transport_type,
+            http_proxy=self._config.http_proxy,
         )
 
-    def close(self):
-        # type: () -> None
+    def close(self) -> None:
         """
         Close down the ServiceBus client.
         All spawned senders, receivers and underlying connection will be shutdown.
@@ -172,7 +189,7 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
         self._handlers.clear()
 
         if self._connection_sharing and self._connection:
-            self._connection.destroy()
+            self._connection.close()
 
     @classmethod
     def from_connection_string(
@@ -215,6 +232,9 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
         :keyword str connection_verify: Path to the custom CA_BUNDLE file of the SSL certificate which is used to
          authenticate the identity of the connection endpoint.
          Default is None in which case `certifi.where()` will be used.
+        :keyword uamqp_transport: Whether to use the `uamqp` library as the underlying transport. The default value is
+         False and the Pure Python AMQP library will be used as the underlying transport.
+        :paramtype uamqp_transport: bool
         :rtype: ~azure.servicebus.ServiceBusClient
 
         .. admonition:: Example:
@@ -245,8 +265,11 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
             **kwargs
         )
 
-    def get_queue_sender(self, queue_name, **kwargs):
-        # type: (str, Any) -> ServiceBusSender
+    def get_queue_sender(
+        self,
+        queue_name: str,
+        **kwargs: Any
+    ) -> ServiceBusSender:
         """Get ServiceBusSender for the specific queue.
 
         :param str queue_name: The path of specific Service Bus Queue the client connects to.
@@ -288,6 +311,7 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
             retry_backoff_max=self._config.retry_backoff_max,
             custom_endpoint_address=self._custom_endpoint_address,
             connection_verify=self._connection_verify,
+            amqp_transport=self._amqp_transport,
             **kwargs
         )
         self._handlers.add(handler)
@@ -402,6 +426,7 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
             prefetch_count=prefetch_count,
             custom_endpoint_address=self._custom_endpoint_address,
             connection_verify=self._connection_verify,
+            amqp_transport=self._amqp_transport,
             **kwargs
         )
         self._handlers.add(handler)
@@ -449,6 +474,7 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
             retry_backoff_max=self._config.retry_backoff_max,
             custom_endpoint_address=self._custom_endpoint_address,
             connection_verify=self._connection_verify,
+            amqp_transport=self._amqp_transport,
             **kwargs
         )
         self._handlers.add(handler)
@@ -562,6 +588,7 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
                 prefetch_count=prefetch_count,
                 custom_endpoint_address=self._custom_endpoint_address,
                 connection_verify=self._connection_verify,
+                amqp_transport=self._amqp_transport,
                 **kwargs
             )
         except ValueError:
@@ -591,6 +618,7 @@ class ServiceBusClient(object): # pylint: disable=client-accepts-api-version-key
                 prefetch_count=prefetch_count,
                 custom_endpoint_address=self._custom_endpoint_address,
                 connection_verify=self._connection_verify,
+                amqp_transport=self._amqp_transport,
                 **kwargs
             )
         self._handlers.add(handler)