+++ sdk/containerregistry/azure-containerregistry/CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Features Added
 
 ### Breaking Changes
+- Added sanity check for manifest size at download time - if manifest is bigger than 4MB, `ValueError` will be thrown.
 
 ### Bugs Fixed
 
 ### Other Changes
+- Changed the digest validation exception type to `DigestValidationError` in `set/get manifest` and `upload/download blob` operations.
 
 ## 1.1.0b4 (2023-04-25)